generate ImportFromLegacy() (#3445)
