+++ models/rawrecords.go
@@ -3,7 +3,6 @@ package models
 import (
 	"fmt"
 
-	"github.com/StackExchange/dnscontrol/v4/pkg/fieldtypes"
 	"github.com/go-acme/lego/v4/log"
 )
 
@@ -52,41 +51,6 @@ func CheckAndFixImport(recs []*RecordConfig, origin string) bool {
 		}
 	}
 	return found
-}
-
-// ImportFromLegacy copies the legacy fields (MxPreference, SrvPort, etc.) to
-// the .Fields structure.  It is the reverse of Seal*().
-func (rc *RecordConfig) ImportFromLegacy(origin string) error {
-
-	if IsTypeLegacy(rc.Type) {
-		// Nothing to convert!
-		return nil
-	}
-
-	switch rc.Type {
-	case "A":
-		ip, err := fieldtypes.ParseIPv4(rc.target)
-		if err != nil {
-			return err
-		}
-		return RecordUpdateFields(rc, A{A: ip}, nil)
-	case "MX":
-		return RecordUpdateFields(rc,
-			MX{Preference: rc.MxPreference, Mx: rc.target},
-			nil,
-		)
-	case "CNAME":
-		return RecordUpdateFields(rc,
-			CNAME{Target: rc.target},
-			nil,
-		)
-	case "SRV":
-		return RecordUpdateFields(rc,
-			SRV{Priority: rc.SrvPriority, Weight: rc.SrvWeight, Port: rc.SrvPort, Target: rc.target},
-			nil,
-		)
-	}
-	panic("Should not happen")
 }
 
 // MustImportFromLegacy is like ImportFromLegacy but panics on error. Use only in tests and init() functions.
